QukiHub/requirements.txt
boto==2.41.0
Django==1.9.7
django-ckeditor==5.0.3
django-pipeline==1.6.8
django-storages-redux==1.3.2
docutils==0.12
jmespath==0.9.0
python-dateutil==2.5.3
six==1.10.0
